feat: citysuggestion 자동스크롤 구현
